Refactor test case submission logic to use a dedicated function for improved readability and maintainability
