+++ openmetadata-ui/src/main/resources/ui/playwright/e2e/Features/TestSuiteMultiPipeline.spec.ts
@@ -19,6 +19,7 @@ import {
   selectAddObservabilityFeature,
 } from '../../utils/dataQuality';
 import { waitForAllLoadersToDisappear } from '../../utils/entity';
+import { submitTestCaseForm } from '../../utils/testCases';
 
 // use the admin user to login
 test.use({ storageState: 'playwright/.auth/admin.json' });
@@ -61,13 +62,7 @@ test(
       await page.getByTestId('test-type').locator('div').click();
       await page.getByTestId('tableColumnCountToEqual').click();
       await page.getByPlaceholder('Enter a Count').fill('13');
-      const createTestCaseResponse = page.waitForResponse(
-        (response) =>
-          response.url().includes('/api/v1/dataQuality/testCases') &&
-          response.request().method() === 'POST'
-      );
-      await page.getByTestId('create-btn').click();
-      await createTestCaseResponse;
+      await submitTestCaseForm(page);
 
       await page.reload();
       await waitForAllLoadersToDisappear(page);
