+++ openmetadata-ui/src/main/resources/ui/playwright/e2e/Pages/DataContracts.spec.ts
@@ -85,6 +85,7 @@ import {
 } from '../../utils/entity';
 import { navigateToPersonaWithPagination } from '../../utils/persona';
 import { settingClick } from '../../utils/sidebar';
+import { submitTestCaseForm } from '../../utils/testCases';
 import { test } from '../fixtures/pages';
 
 // Define entities that support Data Contracts
@@ -448,20 +449,7 @@ test.describe('Data Contracts', () => {
             'Pipeline will only be triggered manually.'
           );
 
-          const pipelineResponse = page.waitForResponse(
-            '/api/v1/services/ingestionPipelines'
-          );
-          const deploy = page.waitForResponse(
-            '/api/v1/services/ingestionPipelines/deploy/*'
-          );
-
-          const testCaseResponse = page.waitForResponse(
-            '/api/v1/dataQuality/testCases'
-          );
-          await page.click('[data-testid="create-btn"]');
-          await testCaseResponse;
-          await pipelineResponse;
-          await deploy;
+          await submitTestCaseForm(page);
 
           await expect(page.getByRole('dialog')).not.toBeVisible();
